zenml/zenml_requirements.txt
zenml==0.12.0
pillow==7.2.0
tensorflow==2.3.0
